pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "homelab-subtitle-service"
version = "0.1.0"
description = "CLI and service for generating subtitles from video files using ffmpeg and Whisper."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Jose Ibañez Ortiz", email = "118691212+Jurel89@users.noreply.github.com" }
]

keywords = ["subtitles", "ffmpeg", "whisper", "media", "homelab"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  # Faster, nicer Whisper implementation
  "faster-whisper>=1.0.0",
  # For batch config files (YAML)
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
]

[project.scripts]
# This is the CLI entrypoint: `subsvc` will call homelab_subs.cli:main
subsvc = "homelab_subs.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/homelab_subs"]
